+++ lib/ferrum/browser.rb
@@ -28,7 +28,7 @@ module Ferrum
                 add_script_tag add_style_tag bypass_csp
                 on position position=
                 playback_rate playback_rate=
-                disable_javascript] => :page
+                disable_javascript set_viewport] => :page
     delegate %i[default_user_agent] => :process
 
     attr_reader :client, :process, :contexts, :options, :window_size, :base_url